prov/sockets: Move MR functionality to a separate source file
