+++ prov/sockets/Makefile.include
@@ -4,31 +4,32 @@ if HAVE_SOCKETS
 
 AM_CPPFLAGS += -I$(top_srcdir)/prov/sockets/include -I$(top_srcdir)/prov/sockets
 
-_sockets_files = \
-	prov/sockets/src/sock_av.c \
-	prov/sockets/src/sock_dom.c \
-	prov/sockets/src/sock_eq.c \
-	prov/sockets/src/sock_cq.c \
-	prov/sockets/src/sock_cntr.c \
-	prov/sockets/src/sock_poll.c \
-	prov/sockets/src/sock_wait.c \
-	prov/sockets/src/sock_ep_rdm.c \
-	prov/sockets/src/sock_ep_dgram.c \
-	prov/sockets/src/sock_ep_msg.c \
-	prov/sockets/src/sock_fabric.c \
-	prov/sockets/src/sock_ep.c \
-	prov/sockets/src/sock_ctx.c \
-	prov/sockets/src/sock_rx_entry.c \
-	prov/sockets/src/sock_progress.c \
-	prov/sockets/src/sock_comm.c \
-	prov/sockets/src/sock_conn.c \
-	prov/sockets/src/sock_msg.c \
-	prov/sockets/src/sock_rma.c \
-	prov/sockets/src/sock_atomic.c \
+_sockets_files =				\
+	prov/sockets/src/sock_av.c		\
+	prov/sockets/src/sock_dom.c		\
+	prov/sockets/src/sock_mr.c		\
+	prov/sockets/src/sock_eq.c		\
+	prov/sockets/src/sock_cq.c		\
+	prov/sockets/src/sock_cntr.c		\
+	prov/sockets/src/sock_poll.c		\
+	prov/sockets/src/sock_wait.c		\
+	prov/sockets/src/sock_ep_rdm.c		\
+	prov/sockets/src/sock_ep_dgram.c	\
+	prov/sockets/src/sock_ep_msg.c		\
+	prov/sockets/src/sock_fabric.c		\
+	prov/sockets/src/sock_ep.c		\
+	prov/sockets/src/sock_ctx.c		\
+	prov/sockets/src/sock_rx_entry.c	\
+	prov/sockets/src/sock_progress.c	\
+	prov/sockets/src/sock_comm.c		\
+	prov/sockets/src/sock_conn.c		\
+	prov/sockets/src/sock_msg.c		\
+	prov/sockets/src/sock_rma.c		\
+	prov/sockets/src/sock_atomic.c		\
 	prov/sockets/src/sock_trigger.c
 
-_sockets_headers = \
-	prov/sockets/include/sock.h \
+_sockets_headers =			\
+	prov/sockets/include/sock.h 	\
 	prov/sockets/include/sock_util.h
 
 if HAVE_SOCKETS_DL
